LayoutTests/fast/canvas/html-in-canvas/drawElementImage-2arg-snapshot-count.html

<!DOCTYPE html><!-- webkit-test-runner [ CanvasDrawElementEnabled=true ] -->
<title>drawElementImage snapshot count tracking (TB1b)</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<canvas id="c" layoutsubtree>
  <div id="a" style="background:red; width:50px; height:50px;"></div>
  <div id="b" style="background:blue; width:50px; height:50px;"></div>
</canvas>
<script>
function nextPaint() {
    return new Promise(r => requestAnimationFrame(() => requestAnimationFrame(r)));
}

promise_test(async () => {
    const canvas = document.getElementById("c");
    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2,
        "two direct children -> two snapshots after paint");
}, "snapshot count reflects direct children after paint");

promise_test(async () => {
    const canvas = document.getElementById("c");
    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2);

    canvas.width = canvas.width + 1;
    assert_equals(internals.canvasChildSnapshotCount(canvas), 0,
        "resize clears snapshots synchronously");

    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2,
        "next frame re-records");
}, "canvas resize clears snapshots, next paint re-records");

promise_test(async () => {
    const canvas = document.getElementById("c");
    await nextPaint();
    assert_equals(internals.canvasChildSnapshotCount(canvas), 2);

    const a = document.getElementById("a");
    canvas.removeChild(a);
    assert_equals(internals.canvasChildSnapshotCount(canvas), 1,
        "removing a child immediately purges its snapshot");
    canvas.appendChild(a); // restore for any later tests in this file
}, "removing a direct child purges its snapshot");
</script>
</body>
